# api/swagger.yaml
openapi: 3.0.0
info:
  title: Books API
  description: This web service offers information on books
  version: 0.1.9
servers:
- url: http://webservice_booksapps.netlify.org
  description: web service published on netlify
paths:
...
  /books/:
    get:
      summary: Returns all books
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    put:
      summary: Create a new book
      requestBody:
        description: Book data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        "200":
          description: OK
  /publishers/:
    get:
      summary: Returns all publishers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publisher'
    put:
      summary: Create a new publisher
      requestBody:
        description: Publisher data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Publisher'
        required: true
      responses:
        "200":
          description: OK
  /authors/:
    get:
      summary: Returns all authors
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
    put:
      summary: Create a new author
      requestBody:
        description: Author data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
        required: true
      responses:
        "200":
          description: OK
components:
  schemas:
    Book:
      type: object
      properties:
        bookId:
          type: string
        title:
          type: string
        edition:
          type: string
        copyright:
          type: string
        language:
          type: string
        pages:
          type: string
        author:
          type: array
          items:
            type: string
        publisher:
          type: string
      example:
        copyright: "2015"
        pages: "1054"
        author: {}
        edition: 4th
        publisher: Publisher1
        language: ENGLISH
        title: Database System
        bookId: book01
    Publisher:
      type: object
      properties:
        publisherId:
          type: string
        name:
          type: string
        country:
          type: string
        founded:
          type: string
        genere:
          type: string
      example:
        name: Mc Graw Hill
        country: US
        founded: "1993"
        genere: Computer Science
        publisherId: publ01
    Author:
      type: object
      properties:
        authorId:
          type: string
        name:
          type: string
        nationality:
          type: string
        birth:
          type: string
        genere:
          type: string
      example:
        name: Diaz
        nationality: Costa Rica
        birth: "1973"
        genere: Computer Science
        authorId: auth01